requirements.txt
evaluate==0.3.0
jsonlines==3.1.0
loguru==0.6.0
numpy==1.19.2
omegaconf==2.2.3
PyYAML==6.0
torch==1.11.0
tqdm==4.62.3
transformers==4.25.1
